+++ src/bitmap/color_plate_scanner.hpp
@@ -243,13 +243,13 @@ namespace Invader {
         void read_unrolled_cubemap(GeneratedBitmapData &generated_bitmap, const ColorPlatePixel *pixels, std::uint32_t width, std::uint32_t height) const;
 
         /**
-         * Read bitmap data that doesn't have a valid color plate
+         * Read bitmap data that doesn't have a valid color plate but still uses blue borders
          * @param generated_bitmap bitmap data to write to (output)
          * @param pixels           pixel input
          * @param width            width of input
          * @param height           height of input
          */
-        void read_non_color_plate(GeneratedBitmapData &generated_bitmap, const ColorPlatePixel *pixels, std::uint32_t width, std::uint32_t height) const;
+        void read_blue_border(GeneratedBitmapData &generated_bitmap, const ColorPlatePixel *pixels, std::uint32_t width, std::uint32_t height) const;
 
         /**
          * Read a single bitmap.